+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Changelog
  * added `IOException` and better error handling in `JsonEncoder::encodeFile()`
    and `JsonDecoder::decodeFile()`
  * `JsonEncoder::encodeFile()` now creates missing directories on demand
+ * `JsonEncoder` now throws an exception on all PHP versions when binary values 
+   are passed
 
 * 1.0.2 (2015-08-11)